chore(api): move acceptance tests to appropriate files
