+++ api/tests/team/acceptance/application/certification-center-membership/certification-center-membership.route.test.js
@@ -95,6 +95,130 @@ describe('Acceptance | Team | Application | Routes | certification-center-member
         expect(_.omit(response.result, 'included')).to.deep.equal(expectedUpdatedCertificationCenterMembership);
       });
     });
+
+    context('Error cases', function () {
+      context('when current user has a member role', function () {
+        it('returns a 403 HTTP error for missing or insufficient permissions', async function () {
+          // given
+          const user = databaseBuilder.factory.buildUser();
+          const certificationCenter = databaseBuilder.factory.buildCertificationCenter();
+          const certifCenterMemberUser = databaseBuilder.factory.buildUser.withCertificationCenterMembership({
+            certificationCenterId: certificationCenter.id,
+            role: 'MEMBER',
+          });
+          const certificationCenterMembership = databaseBuilder.factory.buildCertificationCenterMembership({
+            certificationCenterId: certificationCenter.id,
+            userId: user.id,
+          });
+          await databaseBuilder.commit();
+
+          const request = {
+            method: 'PATCH',
+            url: `/api/certification-centers/${certificationCenter.id}/certification-center-memberships/${certificationCenterMembership.id}`,
+            payload: {
+              id: user.id,
+              data: {
+                type: 'certification-center-memberships',
+                'certification-center-membership-id': certificationCenterMembership.id.toString(),
+                attributes: {
+                  role: 'ADMIN',
+                },
+              },
+            },
+            headers: generateAuthenticatedUserRequestHeaders({ userId: certifCenterMemberUser.id }),
+          };
+
+          // when
+          const response = await server.inject(request);
+
+          // then
+          expect(response.statusCode).to.equal(403);
+          expect(response.result.errors[0].detail).to.equal('Missing or insufficient permissions.');
+        });
+      });
+    });
+
+    context('when certification center membership does not belong to the certification center', function () {
+      it('returns a 403 HTTP error for wrong certification center passed', async function () {
+        // given
+        const user = databaseBuilder.factory.buildUser();
+        const certificationCenter = databaseBuilder.factory.buildCertificationCenter();
+        const certifCenterAdminUser = databaseBuilder.factory.buildUser.withCertificationCenterMembership({
+          certificationCenterId: certificationCenter.id,
+          role: 'ADMIN',
+        });
+
+        const anotherCertificationCenter = databaseBuilder.factory.buildCertificationCenter();
+        const anotherCertificationCenterMembership = databaseBuilder.factory.buildCertificationCenterMembership({
+          certificationCenterId: anotherCertificationCenter.id,
+        });
+        const anotherCertificationCenterMembershipId = anotherCertificationCenterMembership.id;
+        await databaseBuilder.commit();
+
+        const request = {
+          method: 'PATCH',
+          url: `/api/certification-centers/${certificationCenter.id}/certification-center-memberships/${anotherCertificationCenterMembershipId}`,
+          payload: {
+            data: {
+              id: user.id,
+              type: 'certification-center-memberships',
+              attributes: {
+                role: 'ADMIN',
+                'certification-center-membership-id': anotherCertificationCenterMembershipId,
+              },
+            },
+          },
+          headers: generateAuthenticatedUserRequestHeaders({ userId: certifCenterAdminUser.id }),
+        };
+
+        // when
+        //const { statusCode } = await server.inject(request);
+        const response = await server.inject(request);
+
+        // then
+        expect(response.statusCode).to.equal(403);
+        expect(response.result.errors[0].detail).to.equal('Wrong certification center');
+      });
+    });
+
+    context('when certification center membership ID in url is not valid', function () {
+      it('returns a 400 HTTP status code', async function () {
+        // given
+        const user = databaseBuilder.factory.buildUser();
+        const certificationCenter = databaseBuilder.factory.buildCertificationCenter();
+        const certifCenterAdminUser = databaseBuilder.factory.buildUser.withCertificationCenterMembership({
+          certificationCenterId: certificationCenter.id,
+          role: 'ADMIN',
+        });
+        const certificationCenterMembership = databaseBuilder.factory.buildCertificationCenterMembership({
+          certificationCenterId: certificationCenter.id,
+          userId: user.id,
+        });
+        await databaseBuilder.commit();
+
+        const request = {
+          method: 'PATCH',
+          url: `/api/certification-centers/${certificationCenter.id}/certification-center-memberships/${certificationCenterMembership}`,
+          payload: {
+            data: {
+              id: user.id,
+              type: 'certification-center-memberships',
+              attributes: {
+                role: 'ADMIN',
+                'certification-center-membership-id': certificationCenterMembership,
+              },
+            },
+          },
+          headers: generateAuthenticatedUserRequestHeaders({ userId: certifCenterAdminUser.id }),
+        };
+
+        // when
+        const { statusCode } = await server.inject(request);
+
+        // then
+        expect(statusCode).to.equal(400);
+      });
+    });
   });
 
   describe('POST /api/certif/certification-centers/{certificationCenterId}/update-referer', function () {
@@ -178,6 +302,32 @@ describe('Acceptance | Team | Application | Routes | certification-center-member
 
         // then
         expect(statusCode).to.equal(204);
+      });
+    });
+
+    context('when user does not have a valid role', function () {
+      it('returns a 403 HTTP status code', async function () {
+        const userWithoutRole = databaseBuilder.factory.buildUser();
+        const certificationCenter = databaseBuilder.factory.buildCertificationCenter();
+        const user = databaseBuilder.factory.buildUser();
+        const certificationCenterMembership = databaseBuilder.factory.buildCertificationCenterMembership({
+          certificationCenterId: certificationCenter.id,
+          userId: user.id,
+        });
+
+        const request = {
+          method: 'DELETE',
+          url: `/api/certification-center-memberships/${certificationCenterMembership.id}`,
+          headers: generateAuthenticatedUserRequestHeaders({ userId: userWithoutRole.id }),
+        };
+
+        await databaseBuilder.commit();
+
+        // when
+        const { statusCode } = await server.inject(request);
+
+        // then
+        expect(statusCode).to.equal(403);
       });
     });
   });
